fix(app): clear Kotlin compiler warnings
- quadraticBezierTo -> quadraticTo (deprecated alias) in the onboarding shield
- drop redundant .toDouble() on the already-Double strength weight (PathLoss,
  TargetEstimator)
- drop an unnecessary !! after the null-guard (SignalGrid.strongest,
  TrackingSession heading delta)
- opt composeApp into -Xexpect-actual-classes (BackgroundScan expect/actual),
  matching :core, to silence the Beta warning

Build-config (AGP 9->10 project structure + compose.* DSL accessors) is being
handled separately on an isolated branch.

+++ app/core/src/commonMain/kotlin/fyi/blep/core/spatial/PathLossCalibrator.kt
@@ -22,7 +22,7 @@ object PathLossCalibrator {
         for (p in points) {
             val d = max((p.pos - target).length, 0.5)
             val x = log10(d)
-            val w = (p.strength01 + 0.05).toDouble()
+            val w = p.strength01 + 0.05
             wSum += w; sx += w * x; sy += w * p.rssi
             if (x < minX) minX = x; if (x > maxX) maxX = x
         }
@@ -32,7 +32,7 @@ object PathLossCalibrator {
         for (p in points) {
             val d = max((p.pos - target).length, 0.5)
             val x = log10(d) - mx
-            val w = (p.strength01 + 0.05).toDouble()
+            val w = p.strength01 + 0.05
             sxx += w * x * x; sxy += w * x * (p.rssi - my)
         }
         if (sxx < 1e-9) return null
